+++ result/normC.c
@@ -21,20 +21,20 @@
  * @return :     ortho-nomalized Cbn
  * @bug    :     make angle unsteady everytime GPS
  * TODO    :     quaternion or ortho-normalized
- * Arguments    : double C[9]
- *                double Cbn[9]
+ * Arguments    : float C[9]
+ *                float Cbn[9]
  * Return Type  : void
  */
-void normC(double C[9], double Cbn[9])
+void normC(float C[9], float Cbn[9])
 {
-  double dv0[9];
+  float dv0[9];
   int k;
-  double x[3];
-  double y;
+  float x[3];
+  float y;
   int i0;
 
   /* % METHOD web */
-  skew_mat3(*(double (*)[3])&C[3], dv0);
+  skew_mat3(*(float (*)[3])&C[3], dv0);
   for (k = 0; k < 3; k++) {
     C[k] = 0.0;
     for (i0 = 0; i0 < 3; i0++) {
@@ -42,7 +42,7 @@ void normC(double C[9], double Cbn[9])
     }
   }
 
-  power(*(double (*)[3])&C[0], x);
+  power(*(float (*)[3])&C[0], x);
   y = x[0];
   for (k = 0; k < 2; k++) {
     y += x[k + 1];
@@ -53,7 +53,7 @@ void normC(double C[9], double Cbn[9])
     C[k] /= y;
   }
 
-  skew_mat3(*(double (*)[3])&C[6], dv0);
+  skew_mat3(*(float (*)[3])&C[6], dv0);
   for (k = 0; k < 3; k++) {
     C[3 + k] = 0.0;
     for (i0 = 0; i0 < 3; i0++) {
@@ -61,7 +61,7 @@ void normC(double C[9], double Cbn[9])
     }
   }
 
-  power(*(double (*)[3])&C[3], x);
+  power(*(float (*)[3])&C[3], x);
   y = x[0];
   for (k = 0; k < 2; k++) {
     y += x[k + 1];
@@ -72,7 +72,7 @@ void normC(double C[9], double Cbn[9])
     C[3 + k] /= y;
   }
 
-  power(*(double (*)[3])&C[6], x);
+  power(*(float (*)[3])&C[6], x);
   y = x[0];
   for (k = 0; k < 2; k++) {
     y += x[k + 1];
@@ -83,7 +83,7 @@ void normC(double C[9], double Cbn[9])
     C[6 + k] /= y;
   }
 
-  memcpy(&Cbn[0], &C[0], 9U * sizeof(double));
+  memcpy(&Cbn[0], &C[0], 9U * sizeof(float));
 }
 
 /*
